+++ tts_middleware/text/num2word.py
@@ -1,42 +1,21 @@
 import re
 
 from itertools import repeat
-from concurrent.futures import ThreadPoolExecutor
 
 from google.transliteration import transliterate_text
-from tts_middleware.num_to_words.num_to_word import num_to_word
-from dateutil.parser import parse
+from tts_middleware.text.num_to_words.num_to_word import num_to_word
+
 
 MONTH_MAPPING = {1: 'जनवरी', 2: 'फरवरी', 3: 'मार्च', 4: 'अप्रैल', 5: 'मई',
              6: 'जून', 7: 'जुलाई', 8: 'अगस्त', 9: 'सितम्बर', 10: 'अक्टूबर',
              11: 'नवम्बर', 12: 'दिसम्बर'}
+
 
 CHAR_ENG_TO_HIN_MAP = {'a': 'ए', 'b': 'बी', 'c': 'सी', 'd': 'डी', 'e': 'ई', 'f': 'एफ', 
 'g': 'जी', 'h': 'एच', 'i': 'आयी', 'j': 'जे', 'k': 'के', 'l': 'एल', 'm': 'एम', 'n': 'एन', 
 'o': 'ओ', 'p': 'पी', 'q': 'क्यू', 'r': 'आर', 's': 'एस', 't': 'टी', 'u': 'यू', 'v': 'वी', 
 'w': 'डब्ल्यू', 'x': 'एक्स', 'y': 'वायी', 'z': 'जेड'}
 
-def preprocess_text(text, **kwargs):
-    transliterate = kwargs["transliterate"]
-    language_code = kwargs["language_code"]
-    final_text = text
-    if language_code == "hi":
-        sub_sent = re.sub(r',', " , ", text)
-        sub_sent = re.sub(r'\?', " ? ", sub_sent)
-        sub_sent = re.sub(r'।', " । ", sub_sent)
-        sub_sent = re.sub(r'!', " ! ", sub_sent)
-        sub_sent = re.sub(r'\|', " । ", sub_sent)
-
-        with ThreadPoolExecutor(max_workers=8) as executor:
-            inter_text = list(executor.map(process, sub_sent.split(" "), repeat(language_code), repeat(transliterate)))   
-        final_text = " ".join(inter_text).strip()
-
-        sent_ending = r'[\?।|!]'
-        if not re.match(sent_ending, final_text[-1]):
-            final_text += ' ।'
-    final_text = re.compile(r"\s+").sub(" ", final_text).strip()
-
-    return final_text
 
 def parse_date(word, language_code):
     yy, mm, dd = word.split('-')
@@ -60,6 +39,7 @@ def parse_date(word, language_code):
     hi_mm = MONTH_MAPPING[int(mm.strip())]
 
     return " ".join([hi_dd, hi_mm, ", " + hi_yy]) 
+
 
 def process(word, language_code, transliterate):
     final_word = word
